+++ cogment/env_service.py
@@ -19,7 +19,8 @@ import cogment.api.environment_pb2_grpc as env_grpc_api
 import cogment.api.environment_pb2 as env_api
 import cogment.api.common_pb2 as common_api
 
-from cogment.utils import list_versions, logger, INIT_TIMEOUT
+import cogment.utils as utils
+from cogment.utils import logger
 from cogment.session import RecvEvent, RecvMessage, ActorStatus, RecvAction, EventType
 from cogment.session import _InitAck, _EndingAck, _Ending
 from cogment.errors import CogmentError
@@ -410,7 +411,7 @@ class EnvironmentServicer(env_grpc_api.EnvironmentSPServicer):
         trial_id = metadata["trial-id"]
 
         try:
-            init_data = await asyncio.wait_for(self._get_init_data(context, trial_id), INIT_TIMEOUT)
+            init_data = await asyncio.wait_for(self._get_init_data(context, trial_id), utils.INIT_TIMEOUT)
             if init_data is None:
                 return
             key = _trial_key(trial_id, init_data.name)
@@ -438,8 +439,41 @@ class EnvironmentServicer(env_grpc_api.EnvironmentSPServicer):
     # Override
     async def Version(self, request, context):
         try:
-            return list_versions()
+            return utils.list_versions()
 
         except Exception:
             logger.exception("Version")
             raise
+
+    # Override
+    async def Status(self, request, context):
+        reply = common_api.StatusReply()
+
+        try:
+            if len(request.names) == 0:
+                return reply
+
+            # We purposefully don't scan for "*" ahead of time to allow explicit values before
+            all = False
+            for name in request.names:
+                if name == "*":
+                    all = True
+
+                if all or name == "overall_load":
+                    load = utils.machine_load()
+                    if load:
+                        reply.statuses["overall_load"] = str(load)
+                    else:
+                        reply.statuses["overall_load"] = "0"
+
+                if all or name == "nb_sessions":
+                    reply.statuses["nb_sessions"] = str(len(self._sessions))
+
+                if all:
+                    break
+
+        except Exception:
+            logger.exception("Status")
+            raise
+
+        return reply