+++ cogment/hooks_service.py
@@ -14,12 +14,14 @@
 
 import cogment.api.hooks_pb2_grpc as hooks_grpc_api
 import cogment.api.hooks_pb2 as hooks_api
+import cogment.api.common_pb2 as common_api
 
 from cogment.trial import Trial
 from cogment.prehook import PrehookSession
 from cogment.parameters import TrialParameters
 from cogment.errors import CogmentError
-from cogment.utils import list_versions, logger
+import cogment.utils as utils
+from cogment.utils import logger
 
 import asyncio
 
@@ -171,8 +173,41 @@ class PrehookServicer(hooks_grpc_api.TrialHooksSPServicer):
     # Override
     async def Version(self, request, context):
         try:
-            return list_versions()
+            return utils.list_versions()
 
         except Exception:
             logger.exception("Version")
             raise
+
+    # Override
+    async def Status(self, request, context):
+        reply = common_api.StatusReply()
+
+        try:
+            if len(request.names) == 0:
+                return reply
+
+            # We purposefully don't scan for "*" ahead of time to allow explicit values before
+            all = False
+            for name in request.names:
+                if name == "*":
+                    all = True
+
+                if all or name == "overall_load":
+                    load = utils.machine_load()
+                    if load:
+                        reply.statuses["overall_load"] = str(load)
+                    else:
+                        reply.statuses["overall_load"] = "0"
+
+                if all or name == "nb_sessions":
+                    reply.statuses["nb_sessions"] = str(len(self._sessions))
+
+                if all:
+                    break
+
+        except Exception:
+            logger.exception("Status")
+            raise
+
+        return reply